fix: add pinebook pro wifi firmware and local niri
the pinebook pro needs extra wifi firmware that still isn't included in packages, so this adds it.

additionally, the pinebook pro's SoC presently requires a weird config hack to run Niri. this change adds a niri config with just that hack, and also adds local niri config files, in order to make performing this hack as simple as `include`-ing the files.

Lastly, set the Trans Flag Sunset wallpaper as the default Noctalia one, too!